requirements-fixed.txt
numpy==1.24.3
scipy==1.10.1
protobuf==3.20.3
tensorflow==2.13.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
flask==2.3.3
flask-sqlalchemy==3.1.1
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
pyttsx3==2.90
pillow==10.1.0